Extract colour utils to separate module
